chore: Move tutorial popup to new component
